Small refactorings. Javadocs for classes added. New style for emphasized buttons added.
