+++ srcalc/src/main/java/gov/va/med/srcalc/domain/variable/Value.java
@@ -1,6 +1,5 @@
 package gov.va.med.srcalc.domain.variable;
 
-import gov.va.med.srcalc.web.ValueVisitor;
 
 /**
  * Represents the given value of a {@link Variable}.
