+++ srcalc/src/main/webapp/WEB-INF/views/displayResults.jsp
@@ -22,14 +22,14 @@
     </c:forEach>
     </ol>
     <ol>
-    <li><button type="submit" disabled>Sign Calculation</button></li>
+    <li><button class="button-em" type="submit" disabled>Sign Calculation</button></li>
     <li>
     <%-- Add the required patientDfn parameter, preserving the patient from the current calculation. --%>
     <c:url var="newCalcUrl" value="/newCalc"><c:param name="patientDfn" value="${calculation.patient.dfn}"/></c:url>
-    <a class="btn-link" href="${newCalcUrl}">Start New Calculation</a></li>
+    <a href="${newCalcUrl}"><button type="button">Start New Calculation</button></a></li>
     <li>
-    	<c:url var="enterVarsUrl" value="/enterVars"><c:param name="patientDfn" value="${calculation.patient.dfn}"/></c:url>
-    	<a class="btn-link" href="${enterVarsUrl}"><button type="submit">Return to Variable Input Form</button></a>
+    	<c:url var="enterVarsUrl" value="/enterVars"/>
+    	<a href="${enterVarsUrl}"><button type="button">Return to Variable Input Form</button></a>
     </li>
     </ol>
     </div>
